client/requirements.txt
PyJWT
flask
python-dotenv
redis
Flask-Session